% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_forest_plot.R
\name{report_p}
\alias{report_p}
\title{report_p}
\usage{
report_p(p, digits = 3)
}
\arguments{
\item{p}{double required}

\item{digits}{significant digits}
}
\value{
p_string a string
}
\description{
report_p
}
\examples{
report_p(0.02018) # returns "p = .020"
report_p(0.00028) # returns "p < .001"
}
